fix for Go 1.10 (#34)
* fix for Go 1.10

A couple of fixes so that fsevents builds in Go 1.10. Sadly this breaks
compatibility with Go 1.9 and below, but there doesn't seem to be any
way to avoid that.

- the `kFSEventStreamEventIdSinceNow` type is now correctly unsigned, so
the the need to convert from a signed integer is no longer needed.

- nil must be changed to 0 in a few places now.

- We must cast to the correct C type in a few places.

* add wrap_deprecated.go for Go 1.9 and older

There were backwards incompatible changes to cgo in Go 1.10:

- https://golang.org/doc/go1.10#cgo

* Update travis.yml

Test with Go 1.8, 1.9, 1.10 and latest stable (in case we forget to
update this later). sw_vers is only present on the osx_image builds so
make the output for it conditional.

* fix to work with Go 1.9.2+