fsevents: improve documentation
This PR adds comprehensive doc comments to all exported
constants, mostly derived from the Apple documentation
but with changes to make them directly relevant to this package's API.

It also uses C constants directly rather than relying on an implicit
association between `1<<iota` and the actual enum values.

It also unexports `GetStreamRefPaths`, `GetStreamRefDescription`
and `GetStreamRefDeviceID`  because none of them are usable
with this package's API because there is no way to get an
`FSEventStreamRef` value because it's hidden inside an unexported
field inside the `EventStream` struct.

Also unexport the `CFRunLoopRef` because it's similarly
not useful or usable.

Other than the above unexports, the API remains identical.

Unfortunately I am unable to get the tests to pass, which
makes me think that something is wrong somewhere, but
that applies to the original code too, not just the code with
these changes applied.